Support Card Message for Dodo
Support Card message for pokemon.This update provides the picture information of the ball and props.For some reason,the information of the pokemon picture needs to be written by the user  and the path of the file should be filled in the setting of "CardImagePath".
If you don't know how to make a suitable dodo image file method and the correct format, you can try to contact me.